Clean up implementation
