+++ multi_llm_chatbot_backend/app/api/routes/__init__.py
@@ -6,6 +6,7 @@ from .provider import router as provider_router
 from .debug import router as debug_router
 from .root import router as root_router
 from .phd_canvas import router as canvas_router
+from .voice import router as voice_router
 
 router = APIRouter()
 router.include_router(chat_router)
@@ -15,3 +16,4 @@ router.include_router(provider_router)
 router.include_router(debug_router)
 router.include_router(root_router)
 router.include_router(canvas_router)
+router.include_router(voice_router, prefix="/api")
