+++ multi_llm_chatbot_backend/app/api/routes/voice.py
@@ -1,47 +1,19 @@
-# NEON AI (TM) SOFTWARE, Software Development Kit & Application Framework
-# All Rights Reserved 2008-2025
-# Licensed under the BSD 3-Clause License
-# https://opensource.org/licenses/BSD-3-Clause
-#
-# Copyright (c) 2008-2025, Neongecko.com Inc.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-# 1. Redistributions of source code must retain the above copyright notice,
-#    this list of conditions and the following disclaimer.
-# 2. Redistributions in binary form must reproduce the above copyright notice,
-#    this list of conditions and the following disclaimer in the documentation
-#    and/or other materials provided with the distribution.
-# 3. Neither the name of the copyright holder nor the names of its contributors
-#    may be used to endorse or promote products derived from this software
-#    without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-"""Voice service health & wake-up management.
-
-Provides status checks and warm-up pings for the external TTS (Coqui) and
-STT (Whisper) services, which may sleep when idle and take time to wake.
-"""
-
 import asyncio
+import html as html_module
 import logging
+import re
+import struct
+import subprocess
+import tempfile
 import time
-from typing import Any, Dict, Optional
+from pathlib import Path
+from typing import Any, Dict, List, Optional
+from urllib.parse import quote
 
 import httpx
-
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, File, HTTPException, UploadFile
+from fastapi.responses import Response
+from pydantic import BaseModel
 
 from app.core.auth import get_current_active_user
 from app.models.user import User
@@ -53,13 +25,24 @@ router = APIRouter()
 COQUI_BASE = "https://coqui.neonaiservices.com"
 WHISPER_BASE = "https://whisper.neonaiservices.com"
 
+HTTP_TIMEOUT = 120
 PROBE_TIMEOUT = 12
 CACHE_TTL = 120
+# Coqui uses GET /synthesize/{text}; keep chunks small for URL limits.
+MAX_SYNTH_CHUNK = 200
 
 _status_cache: Dict[str, Any] = {
     "tts": {"ready": False, "checked_at": 0.0},
     "stt": {"ready": False, "checked_at": 0.0},
 }
+
+_SECTION_HEADERS = re.compile(
+    r"\b(Thought|What to do|Next step)\s*[:.]?\s*", flags=re.IGNORECASE
+)
+
+
+class TTSRequest(BaseModel):
+    text: str
 
 
 async def _probe_tts() -> bool:
@@ -87,7 +70,6 @@ async def _probe_stt() -> bool:
 
 
 def _cached_ready(service: str) -> Optional[bool]:
-    """Return cached readiness if still fresh, else None (unknown)."""
     entry = _status_cache[service]
     if time.time() - entry["checked_at"] < CACHE_TTL:
         return entry["ready"]
@@ -95,15 +77,120 @@ def _cached_ready(service: str) -> Optional[bool]:
 
 
 async def wake_both() -> None:
-    """Fire probes to both services concurrently (background-safe)."""
     await asyncio.gather(_probe_tts(), _probe_stt(), return_exceptions=True)
+
+
+def _rough_spoken_text(raw: str) -> str:
+    """Light cleanup so model markdown is tolerable for TTS (no extra deps)."""
+    t = raw.strip()
+    t = re.sub(r"```[\s\S]*?```", " ", t)
+    t = re.sub(r"`([^`]+)`", r"\1", t)
+    t = re.sub(r"\[([^\]]+)\]\([^)]+\)", r"\1", t)
+    t = re.sub(r"!\[([^\]]*)\]\([^)]+\)", r"\1", t)
+    t = re.sub(r"^#{1,6}\s+", "", t, flags=re.MULTILINE)
+    t = re.sub(r"[*_]{1,3}", "", t)
+    t = html_module.unescape(t)
+    t = _SECTION_HEADERS.sub(" ", t)
+    t = re.sub(r"\s+", " ", t).strip()
+    return t
+
+
+def _text_chunks(text: str, max_len: int = MAX_SYNTH_CHUNK) -> List[str]:
+    text = text.strip()
+    if not text:
+        return []
+    out: List[str] = []
+    while text:
+        if len(text) <= max_len:
+            out.append(text)
+            break
+        cut = text.rfind(" ", 0, max_len)
+        if cut <= 0:
+            cut = max_len
+        piece = text[:cut].strip()
+        if piece:
+            out.append(piece)
+        text = text[cut:].strip()
+    return out
+
+
+async def _synthesize_one(client: httpx.AsyncClient, chunk: str) -> Optional[bytes]:
+    url = f"{COQUI_BASE}/synthesize/{quote(chunk, safe='')}"
+    try:
+        r = await client.get(url)
+        r.raise_for_status()
+        return r.content
+    except Exception as exc:
+        LOG.warning("TTS chunk failed (%s chars): %s", len(chunk), exc)
+        return None
+
+
+def _concat_wav(segments: List[bytes]) -> bytes:
+    if len(segments) == 1:
+        return segments[0]
+
+    pcm_parts: List[bytes] = []
+    first_header = b""
+
+    for i, seg in enumerate(segments):
+        if len(seg) < 44:
+            continue
+        if i == 0:
+            first_header = seg[:44]
+        data_offset = 44
+        idx = seg.find(b"data")
+        if idx != -1 and idx + 8 <= len(seg):
+            data_offset = idx + 8
+        pcm_parts.append(seg[data_offset:])
+
+    if not pcm_parts or not first_header:
+        return segments[0] if segments else b""
+
+    all_pcm = b"".join(pcm_parts)
+    pcm_len = len(all_pcm)
+
+    header = bytearray(first_header)
+    struct.pack_into("<I", header, 4, pcm_len + 36)
+    struct.pack_into("<I", header, 40, pcm_len)
+    return bytes(header) + all_pcm
+
+
+def _convert_to_wav(audio_bytes: bytes, src_mime: str) -> bytes:
+    with tempfile.TemporaryDirectory() as tmp:
+        ext = "webm" if "webm" in (src_mime or "") else "ogg"
+        src = Path(tmp) / f"in.{ext}"
+        dst = Path(tmp) / "out.wav"
+        src.write_bytes(audio_bytes)
+        result = subprocess.run(
+            [
+                "ffmpeg",
+                "-y",
+                "-i",
+                str(src),
+                "-ar",
+                "16000",
+                "-ac",
+                "1",
+                "-f",
+                "wav",
+                str(dst),
+            ],
+            capture_output=True,
+            timeout=30,
+        )
+        if result.returncode != 0:
+            LOG.warning(
+                "ffmpeg stderr: %s",
+                result.stderr.decode(errors="replace")[-500:],
+            )
+            raise RuntimeError("ffmpeg conversion failed")
+        return dst.read_bytes()
 
 
 @router.get("/voice/status")
 async def voice_status(
     current_user: User = Depends(get_current_active_user),
 ) -> Dict[str, bool]:
-    """Return cached readiness or probe fresh."""
     tts_ready = _cached_ready("tts")
     stt_ready = _cached_ready("stt")
 
@@ -119,6 +206,97 @@ async def voice_status(
 async def voice_wake(
     current_user: User = Depends(get_current_active_user),
 ) -> Dict[str, str]:
-    """Kick off warm-up pings for both services and return immediately."""
     asyncio.create_task(wake_both())
     return {"status": "waking"}
+
+
+@router.post("/voice/transcribe")
+async def transcribe_audio(
+    audio: UploadFile = File(...),
+    current_user: User = Depends(get_current_active_user),
+) -> Dict[str, str]:
+    contents = await audio.read()
+    if not contents:
+        return {"text": ""}
+
+    mime = audio.content_type or "audio/webm"
+    LOG.info("STT: received %s bytes (%s)", len(contents), mime)
+
+    need_convert = "wav" not in mime.lower()
+    if need_convert:
+        try:
+            loop = asyncio.get_running_loop()
+            wav_bytes = await loop.run_in_executor(
+                None, _convert_to_wav, contents, mime
+            )
+            LOG.info("STT: converted to WAV (%s bytes)", len(wav_bytes))
+        except Exception as e:
+            LOG.error("STT conversion error: %s", e)
+            raise HTTPException(status_code=500, detail="Audio conversion failed")
+    else:
+        wav_bytes = contents
+
+    try:
+        async with httpx.AsyncClient(timeout=HTTP_TIMEOUT) as client:
+            resp = await client.post(
+                f"{WHISPER_BASE}/stt",
+                content=wav_bytes,
+                headers={"Content-Type": "audio/wav"},
+            )
+            resp.raise_for_status()
+
+            _status_cache["stt"] = {"ready": True, "checked_at": time.time()}
+
+            text = resp.text.strip().strip('"')
+            LOG.info("STT result: %r", text[:100])
+            return {"text": text}
+    except httpx.TimeoutException:
+        raise HTTPException(status_code=504, detail="STT service timed out")
+    except Exception as e:
+        LOG.error("STT proxy error: %s", e)
+        raise HTTPException(status_code=502, detail="STT service unavailable")
+
+
+@router.post("/voice/tts")
+async def text_to_speech(
+    req: TTSRequest,
+    current_user: User = Depends(get_current_active_user),
+) -> Response:
+    raw = req.text.strip()
+    if not raw:
+        raise HTTPException(status_code=400, detail="Text is required")
+
+    if len(raw) > 5000:
+        raw = raw[:5000]
+
+    spoken = _rough_spoken_text(raw)
+    chunks = _text_chunks(spoken)
+    LOG.info(
+        "TTS: %s chunk(s) from %s input chars → %s spoken chars",
+        len(chunks),
+        len(req.text),
+        len(spoken),
+    )
+
+    try:
+        async with httpx.AsyncClient(timeout=HTTP_TIMEOUT) as client:
+            results = await asyncio.gather(
+                *[_synthesize_one(client, c) for c in chunks]
+            )
+            wav_segments = [r for r in results if r and len(r) > 44]
+
+            if not wav_segments:
+                raise HTTPException(
+                    status_code=502, detail="TTS synthesis failed for all chunks"
+                )
+
+            combined = _concat_wav(wav_segments)
+            _status_cache["tts"] = {"ready": True, "checked_at": time.time()}
+            return Response(content=combined, media_type="audio/wav")
+    except HTTPException:
+        raise
+    except httpx.TimeoutException:
+        raise HTTPException(status_code=504, detail="TTS service timed out")
+    except Exception as e:
+        LOG.error("TTS proxy error: %s", e)
+        raise HTTPException(status_code=502, detail="TTS service unavailable")
